Split Menu class into separate module
